Implement `rem(float, float, RoundNearest)` in Julia
One of the few remaining uses of openlibm is implementing `rem` with
`RoundNearest` for `Float32` and `Float64`. This commit translates the
msun libm implementations of `__ieee754_remainder` and
`___ieee754_remainderf` to Julia to avoid relying on openlibm.

The implementations of the `Float32` and `Float64` methods are quite
similar and could be combined with a bit more work.